feat(.storybook): use react-storybook-addon-codesandbox types directly for global SB Parameters extension
