+++ packages/react-components/react-storybook-addon-codesandbox/src/types.ts
@@ -1,16 +1,8 @@
 import type { StoryContext as StoryContextOrigin, Parameters } from '@storybook/addons';
-
-export interface ParametersConfig {
-  optionalDependencies?: Record<string, string>;
-  requiredDependencies?: Record<string, string>;
-  provider: 'codesandbox-cloud' | 'codesandbox-browser' | 'stackblitz-cloud';
-  bundler: 'vite' | 'cra';
-}
-
-export interface ParametersExtension {
-  exportToSandbox?: ParametersConfig;
-}
+import type { ParametersExtension } from './public-types';
 
 export interface StoryContext extends StoryContextOrigin {
   parameters: Parameters & ParametersExtension;
 }
+
+export type { ParametersExtension };
