+++ typings/storybook__addons/index.d.ts
@@ -57,23 +57,6 @@ declare module '@storybook/addons' {
        */
       inlineStories?: boolean;
     };
-    /**
-     * @see https://github.com/microsoft/fluentui-storybook-addons
-     */
-    exportToSandbox?: AddonExportToSandboxParameters;
-  }
-
-  interface AddonExportToSandboxParameters {
-    /**
-     * Dependencies that should be included with every story
-     */
-    requiredDependencies?: Record<string, string>;
-    /**
-     * Dependencies that should be included if the story code contains it
-     */
-    optionalDependencies?: Record<string, string>;
-    provider: 'codesandbox-cloud' | 'codesandbox-browser' | 'stackblitz-cloud';
-    bundler: 'vite' | 'cra';
   }
 
   interface ControlsParameters {
